+++ src/Biome.cpp
@@ -7,12 +7,12 @@
 
 
 
-bool Biome::fillRegions(uint16_t biomeType)
+bool Biome::fillRegions(uint32_t biomeType)
 {
     
-    for(uint16_t x = 0; x < regionMap.size(); ++x)
+    for(uint32_t x = 0; x < regionMap.size(); ++x)
     {
-        for(uint16_t y = 0; y < regionMap[0].size(); ++y)
+        for(uint32_t y = 0; y < regionMap[0].size(); ++y)
         {
             regionMap[x][y].getRegionData().setRegionTileType(biomeType);
             //float elevation = getBiomeData().getElevation();
@@ -23,7 +23,7 @@ bool Biome::fillRegions(uint16_t biomeType)
         
 }
 
-Tile Biome::getTile(uint16_t xPos, uint16_t yPos)
+Tile Biome::getTile(uint32_t xPos, uint32_t yPos)
 {
     Tile tileDisplay = regionMap[xPos][yPos].getRegionData().getTile();
